Summarize the names of collapsed nodes respecting their original order.
This change makes the names of the collapsed nodes appear in order
in their summary string representation.

Before, we used to put them together with a set, so the order was
not preserved. Using a list instead was no good, since for big
trees the speed difference to remove duplicates was tremendous
(~100x slower).

Ideally, we would have used something similar to an OrderedSet in
python, but there is no such thing in the standard library. Instead
we now use an OrderedDict with keys the names of the nodes we want
and values set to None.

They work like a charm :)