Fix default distance for root and non-root nodes.
This was a mess. When creating a node, it had no parent assigned yet, so it would
use as dist DEFAULT_DISTANCE_ROOT, independent of it being assigned a parent
later on.

All the conditionals about self.name, self.dist and self.support were checking
things that could never happen, because all of them have accessors that return
a default value.

We have now moved the node default settings to newick.pyx, since Jaime
thinks that is where they make more sense (I am not so sure, I actually think
that any property not set would be better remain unset, and trying to get
it would either rise an exception or return None). In any case, they are:

  DEFAULT_DIST
  DEFAULT_DIST_ROOT
  DEFAULT_SUPPORT
  DEFAULT_NAME