tree.pyx: Finished __str__() and get_ascii() with full functionality.
The tree __str__() function now calls get_ascii() with certain parameters,
so that the representation is compatible with the old one that ETE had
(but with nicer output!).

The tree get_ascii() function is now also compatible with the old one, and
in addition to all the previous arguments (which still mean the same), now
one can fine-tweak the output by using the padding parameters: px, py, px0.

Also, with waterfall=True it will use the very compact waterfall representation
that we had before with to_str().