Fix reading dist or support from extended properties as a string.
When the dist or support are read from the extended properties, like in:
  (a[&&NHX:common_name=:dist=0.0]);
they were set as strings, and that would cause problems in many other places.

This commit fixes that. Calling the setters for self.dist and self.support
calls the float(value) and catches the exception if there is any.