Add standard dirs ETE_{DATA,CONFIG,CACHE}_HOME, use them for dump files in gtdbquery, and remove dump files.
We want to move towards using XDG-compliant directories, and avoid having data files
in the main repository (specially if they are big).

The current solution using wget, and the points in which it is called, may need to
be reviewed. But the general idea should be correct, and we will change more
into that direction.