Fix not showing popup_prop_keys when using t.explore().
The changes in commit d3039c5ca55dde6 allowed passing popup_prop_keys in
the POST request when adding a tree to a server, but broke the behavior
of t.explore() to get correctly the popup_prop_keys.

This change fixes it, and simplifies the use of popup_prop_keys in general.

The expected behavior is this: all the properties present in MIN_POPUP_PROP_KEYS
will *always* be present, and in addition to those, the ones that the user
passes as popup_prop_keys too.

Thanks Claudia for finding the problem!