popup_prop_keys -> popup_props and change semantics.
We can use a shorter name, popup_props, which will be the properties shown
in the popup for the node.

Also, if the user does not specify anything special in t.explore(), we
want to have all the properties shown. And if popup_props=[..] is passed,
we want to show only the basic properties (name, dist, ...) and the ones
that are specified.

This commit implements it all.