In get_ascii(), change attributes->properties. And tests.
This uses the same term ("properties") as in the rest of the code.

What we call now (and have been calling for a while) "properties", used
to be called "features", and "attributes" in other parts. That was
confusing, and this change is part of an effort to make it all more
uniform (and hopefully clear).