Fix include_properties and exclude_properties when adding a tree with a POST.
To send a tree with an http POST, one should have the include_properties
and exclude_properties properly formatted as lists. But it seems that
people do not easily do that, so instead we now accept them as comma-separated
list of strings, and split it manually when necessary.

A bit ugly, but better than the alternative.