treematcher.py: optimizations, comments and simplification.
The optimization is to test for the length of the children first.

We also allow for patterns that do not end in ";", like:
  ("len(ch) > 2","")"d >= 0.5"

We also let empty nodes (which unfortunately need to be quoted as "", due
to the newick parser). They are equivalent to no condition to check in
that node.