server.py: Add more bottle endpoints and many corrections.
Among the changes, I removed the endpoints:
  /leaves_info
  /descendants_info
which were not working from before I started with the code.

I also fixed /info -> /selection/info

There are many little fixes and improvements too (removing a branch correctly
updates the sizes, for example). But at this stage we are just in the middle
of preparing the move to bottle.