New implementation of to_ultrametric() and get_midpoint_outgroup().
Some changes for consistency among all the functions in ete:
  convert_to_ultrametric() -> to_ultrametric()  (like to_str(), etc.)
  topology_only -> topological  (more accurate and shorter, and used elsewhere)

There are many cosmetic changes too. The tests are adapted, some examples updated,
and a minor bug corrected in get_midpoint_outgroup() too (when called from a node
that was not the root, if the root was the midpoint it would return a wrong node).