Remove unused file_extract().
We were using this when playing with heuristics to know if a string
was a newick or the name of a file. But now the Tree() constructor
does not try to guess. If passed a string, it should be a newick.
Otherwise, one should initialize with Tree(open(...)). Which made
the need for file_extract() obsolete.