Fix importing of faces from smartview.
The faces were being imported from treeview, and if it was
not installed, they would not work when doing things like:

  from ete4 import TextFace

This is crazy! The faces should not be duplicated in treeview and
smartview.

This commit adds a partial fix. In the future there should not be
all these "from .... import *", and the Faces should not be duplicated,
etc., hopefully.