tree.pyx: Add method t.get_prop() that searches in attributes and props.
Normally, when looking at a property in a tree, on wants to use
  t.props
which is the dictionary with all the information.

But sometimes one wants to access an attribute, like for example
in a PhyloTree with t.species (which calls a getter/setter
function).

Since this happens often, this new function provides easy access
for it. It behaves roughly as d.get() for a dictionary.

This commit includes some code rewriten to use that function.