spoverlap.py: Fix undefined call to node distance in get_evol_events_from_leaf().
The function called at a certain point:
  d += current.get_distance(current, leaf)
and that could give an error, trying to add a float to None.

Also, that "branch control" has not been used in a long time and
important parts were commented out.

This commit removes all legacy unused parts of that "branch control",
avoiding that error. Also, it simplifies a few things.