Remove unused tests from ete4/smartview/tests.
Those tests come from the old minimalist version, and test
some things that it would be good for us to test now, and
others that make no sense here (like the access to the
database of users, etc.).

In any case, they are not being used and cannot even be run!

So they are cruft. And now, gone.

It would be very nice to add many of those tests to our
suite of tests in the package /tests directory.