t.set_outgroup(node) works as Jaime expects for node child of root.
Even if the node was already a child of root, we want set_outgroup() to
make the distances equal, and resolve politomies if there were any.

The added tests check for this. And actually the code in set_outgroup() is
now cleaner -- no need to consider any special case, and also fixes copying
the properties of the old root, which could only be node properties (not
branch properties), and so we want to keep them in the root.