Move entry_points to pyproject.toml and set cython language_level=3.
When doing:
  python setup.py develop
I got warning about not setting language_level for cythonize, and setting it
automatically to 2. That is not right!

This commit thus reverts partially da330a1b3f, so we set language_level=3 in
cythonize.

Also, I was getting warnings related to the entry_levels being in setup.py
instead of pyproject.toml. This commit also changes those, following:
https://packaging.python.org/en/latest/specifications/declaring-project-metadata/#entry-points