New version of populate(), with random model selection and distance and support functions.
This version incorporates ideas from https://github.com/etetoolkit/ete/pull/714

We can now select model='yule' or 'uniform'.

And we have more flexibility with how the distances and supports are
added: we can add either random distances and/or random supports,
drawn from whatever distribution we choose, by passing the appropriate
functions as arguments (instead of the tuples of (min,max) that we had
for uniform distributions of both before).

Also, a significant clean up of the code, and updated the tests.