server.py: Fix reloading of trees.
Since commit 13ecff34755ecbb8c, whenever the tree was reloaded
(for example when changing to ultrametric and back, or waiting
for a long time), the tree would not load correctly.

This is in part because the full thing about where data is kept
tracked is a mess.

Anyway, this commit fixes it for the moment.