server.py: Fail gracefully when serializing/deserializing a tree.
Currently, t.explore() will dump the tree to /tmp/*tid*.pickle, and use it
if to "undo" an ultrametric visualization (by reloading the original tree
from disk). Sometimes the trees cannot be pickled, and this was not considered.

Now at least it fails gracefully. But I think we would be better by not dumping
the tree at all, and not allowing the visualization of trees to modify the
tree itself and pretend that it is not. Editing the tree should be explicit.