newick.pyx: Fix overflow when reading very big trees.
When reading a tree from a file with more than 5 GB or so, using
the position as an int would overflow. It happened for example
with K01181_GTDB_rooted.faa.alg_annotated.nw.

Changing int -> long (in cython) for the position fixes the problem,
and has no appreciable bad impact on performance.