Lift `sqlite3`'s `check_same_thread` flag to `NCBITaxa.__init__()`
Using this object in a multithreaded environment ( ex. `concurrent.futures`) results in the following error: 
`SQLite objects created in a thread can only be used in that same thread. The object was created in thread id 131662279252864 and this is thread id 131661501028032.`

Perhaps this is to be expected of sqlite, but given that the taxonomy db itself is readonly ( i assume?) after creation, i think this flag should be exposed.