Get new rendering from the tree-explorer repository.
This is a huge change.

We simplify a lot the use of Faces and Layouts.

The nodes (Tree) do not store the faces anymore. Instead, the node_decos()
function (and {tree,collapsed}_decos(), {node,tree,collapsed}_style())
return a list of Decorations, which contain a Face and its location.

This is a functional interface. It is also simplified in that we do
not need to do complex computations to find the size of the drawn
elements. We also make that the Faces do not need to know in which
position they are going to be. Instead, many drawing functions
return the total size that they occupied, in addition to the basic
drawing elements that they have generated.

We are not on par with the old functionality, though. In particular,
many faces and layouts are missing, and so is the drawing done
with pixi. It is unclear that we will need it, but the idea is to
add in steps all that is needed.

On the other hand, we have new functionalities that did not exist
before, like the outline of collapsed nodes showing the full tree
skeleton. Not to mention there is much less code and it is faster!