draw.js: Make fix_text_orientations() not async.
Why was it async? Can it be too slow or something? Maybe it is
like that from an old time when the function was slower and we
maybe wanted to show some stuff on the window at least?

In any case, it seems not to be appropriate to have it anymore.