tree.pyx: Restore the parts necessary for the treeview module to work.
This includes just add_face(), set_style(), and the property img_style.

Instead of adding an attribute in the Tree class (like
"cdef public object _img_style"), we add the "_img_style" to the
props dictionary, and only in case it is needed. This way we keep
our Tree objects being small, and cleaner (an improvement over the
old way, in my opinion).

Also, we added those only for the old "legacy" treeview (the qt
painting), not for the smartview (the web painting). The smartview
uses the new simplified and cool Layout class and so on, with all the
information about how to represent the nodes outside of the nodes
themselves (so, outside of the Tree class).