Internally cache the decorations for the nodes, and rename layout functions.
This commit changes the names of the functions in a Layout.

For example, instead of having node_style() and node_decos(), we have
now draw_node() that can produce both decorations and styles. The
styles are dictionaries, that we will combine later, by the way.

Also, we cache by default the output of draw_node(), so the
decorations that correspond to a node will be stored and reused (using
lru_cache). This can be controlled with the argument cache_size when
creating a Layout. For example, Layout(..., cache_size=0) will not
cache anything.