newick.pyx: More uniform way managing the newick text.
Basically cosmetics, but also we are more ready to try reading
the different parts of the tree in different Threads, since the
way the functions access the newick would make it easier (always
with the full text and a position where we are).

The speed is basically the same as before. The code should be a bit
more clear.

Also, we are fixing the behavior of certain formats that required
certain fields to be present. Before this change, if there was no
content at all for a node, they would not have been checked.

To do it properly, we set the root node as an exception where those
values do not have to be required.