Add argument is_leaf_fn to explore().
This allows us to specify programatically which nodes to collapse
when exploring the tree.

For example:

  t.explore(is_leaf_fn=lambda node: node.name.startswith("bact"))

This used to exist as a property called "draw_descendants" that you
could assign in a node style (and set with node.set_style()).

Now, it is passed as a function to the explore() function directly,
similar to how is_leaf_fn is used in other parts of the code.