Allow setting style for hz/vt-lines, add new drawing commands, and fixes.
The new drawing commands are:
  draw_hz_line()
  draw_vt_line()

With those, we can control the style of the drawn elements from either
the tree_style, the styles set in the nodes (in the draw_node() of the
layout), or the gui too.

This allows us to simplify the drawing lines command too.

We now change styles coming from the backend by adding the "ns_"
prefix (from "node style", since they refer to the styles defined in
node_styles of the tree style), except if they start with "label_"
(because those styles are managed elsewhere).

The add_style() function is more straightforward. The combination of
it with add_ns_prefix() allows us to mix freely the styles coming
from any source.

This makes the whole style assignment composable, and more
understandable. It was kind of a mess before!