DEFAULT_LAYOUT -> BASIC_LAYOUT
This renaming makes it a bit more informative.

This is for the layout that includes names to the right of the leaf
nodes, distances on the top of the branch, and support on the bottom.

Other ideas that we considered were NAMES_LAYOUT, NAME_DIST_LAYOUT,
NAME_DIST_SUPPORT_LAYOUT.