zoom.js: Fix when zooming out in certain cases.
When zooming out, when the selected node is bigger than the screen,
it was not zooming out, but in!

This had to do with the shape of the sigmoid, which was not an
increasing function from a smaller number than 1, then 1 at 0, then
up to some maximum value where it platoes.