Fix SQL calls with wrong quotation for string literals.
Many SQL statements were written as:
  SELECT ... FROM ... WHERE ... IN ("x","y",...)
and they should have been:
  SELECT ... FROM ... WHERE ... IN ('x','y',...)

Sqlite used to be fine with that, but the new version is more strict about it.
So we needed to update a ton of wrongly quoted strings.

See also https://www.sqlite.org/quirks.html#double_quoted_string_literals_are_accepted

Thanks Ana for spotting this in ete4 and Arnaud in ete3!