tree.pyx: Allow easier typing in t.common_ancestor().
This works a bit more like in ete3. That is, instead of calling:
  t.common_ancestor([n1, n2])
we can now call
  t.common_ancestor(n1, n2)

The old behavior (always passing a list, or a set, etc.) still works
the same.

My original concern that we might want to pass other arguments to
the common_ancestor() function has disappeared, since we have not
changed this function in a long time.

And we can always use the one in operations.pyx, which is more strict
and still requires a list of nodes (and not just node names, or
anything else).