qt_render.py: Fix assigning distance for root node.
When we render a tree, the distance that we use for the nodes that
do not have it defined should be 0 for the root, and 1 for the rest.
We did not cover correctly the case for the root (dist=0 if undefined).

In ete3, all nodes had automatically a dist, and the root had 0. But
in ete4 the nodes do not necessarily have a dist associated (which is
better in many ways). But when we draw it, we need to use some distance
if it has not been defined before, and we were doing it wrong for the
root node.

Thanks Walker Azam for spotting the problem!