Cosmetic changes and bump version to 4.1.1.
The changes are just to follow the toml format more closely.

The main point of this commit is to synchronize the version information
that appears in different files, and make sure that we have a higher
number than any previous one (since we used to have a 4.1.0-beta), so
it is clear that this is the latest one, and we are out of beta.