Add ops.average_distance() and t.average_distance().
This function computes the average distance between a given leaf and
all the selected leaves. Or an average for all selected leaves if no
leaf given.

It does the computation in an efficient way, taking only n*log(n)
instead of n^2 steps (where n is the number of leaves).