tree.pyx: Do some reasonable guess for paths in Tree init.
The idea is to allow for things like:

  Tree(str_that_is_a_path)

instead of having to do

  Tree(open(str_that_is_a_path))

while at the same time, allowing to specify if str_... has to be
interpreted as a path, or as raw data.

The way we do it is, if we want to specify what it is, we use:

  Tree(str_..., parser='data-...')  # or parser='file-...'

So now we can use the init function as before, but for most cases, we
do not need to use the open(...) part, and we can always force ete to
interpret the argument as a path or as data if we really wanted.

It is implemented in a way that should be easy to extend to formats
other than newick, and trying not to complicate the __init__ function
too much.

Also, in newick.pyx we add string parsers corresponding to the numbers, so:

  Tree(str_data, parser=1)

can be used while forcing str_data to be interpreted as a file or raw
data, by doing things like:

  Tree(str_data, parser='data-1')  # or parser='file-1'

I am not going to document this elsewhere for the moment because I am
not convinced it is elegant enough (may be more confusing than useful).