Remove all guesswork from Tree.__init__ and put it in parser/extract.py
This way we clean up Tree.__init__, and also extend the guessing
capabilities when initializing a Tree.

Not all corner cases are covered, but at least it provides quite a bit
of good guesses so loading trees is very convenient, at the same time
as having mechanisms to force to use files or data (with parser optionally
starting with "file-" or "data-").

And it is all in one place, that can be simplified anytime as explained in
the comment at the beginning of the extract_data_parser() function.