indent.py: Nicer error message on badly formated files.
If we are processing an indented file and the jump in indentation
is not consistent, the assertion error will be more useful now.

This happens for example with the kind of indented trees that you
may get from:
  systemctl status
when in the same group of processes there are some with 3 digits and
others with 4, because of the way it indents those lines.