operations.pyx: distance_matrix() optimization.
In this case, creating a generator expression instead of a list
has too much overload for what normally is a list of 2 elements.

Using lists directly instead of generators makes the code 50% faster!
(for example, from 48 ms to 33 ms). This is because we were doing this
unoptimized operation so many times in this function.