tree.pyx: Remove unused and slow sisters() generator.
I introduced this generator a few commits ago (a05165ba), but it is
not used and does not make much sense since generators for so few
elements have more overhead than gains. And it makes more sense to
have the old get_sisters() generate the list directly (faster and
slightly simpler).