operations.pyx: Big optimization in populate_yule() for big trees.
The algorithm we were using was very slow for sizes > 1M nodes or so,
because doing pop()s on a list is very slow, O(N).

With this change, in each iteration we replace the old leaf by one of
the news, and grow the list with append(), which is *much* faster.

For example, for a size of 1M leaves in my laptop it takes:

  48 s (old version)
  10 s (new version)

and the new version actually scales linearly (2M leaves takes ~ 20 s),
which the old version did not.