faces.py: Prepare all faces for dy=0 to mean no limits on dy.
This is the same as the convention we had before for dx.

So if now a faces is asked to be drawn with an available size of dx=0
or dy=0, it means there is no limit in those sizes (and typically the
resulting size results from a maximum of other kind, like the font
size to use or so).

This is also in preparation to use arbitrary faces drawn as headers
too, which I plan to implement soon.