average_distance() takes now a weight_fn instead of a selector.
If we want to calculate the average distance, but weighting leaves
differently, we can now pass a weight_fn as an argument.

The previous case, where we passed a selector function, is a special
case now. For example, instead of the old:

  t.average_distance(selector=lambda node: node.name in names)

we would now do

  t.average_distance(weight_fn=lambda node: 1 if node.name in names else 0)

But in addition to that, we can give arbitrary weights to leaves, and
get meaningful results too.