operations.pyx: Minor fix for get_distance_fn() for root node.
We should not really ask for the distance of the root wherever we are
using this function, but just in case, if we do it in the topological
mode, it should be 0, not 1.

No algorithm was misbehaving because of this (as far as I know and
have seen), but it could have bitten us at some point in the future.
Now it should be safe.