explorer.py: Add POST /load endpoint to load a new tree from a file.
In preparation for using ete as a backend for eggnog7, this endpoint
now allows you to send a POST request to /load with a json:

  [name, path to file, parser name, list of layout names]

This will load and add the tree to the ones that the server is
keeping, and it will be available with all the functions at the
endpoint

  /trees/<name>/...

(In particular, for drawing with GET /trees/<name>/draw...)

The path must be a valid path with a tree in the system where ete is
running. The parser can be 'newick', a number, 'nexus', 'indent',
'ete', etc.

The layouts must have been loaded for some tree before, so that they
are available when we try to load the new tree. For example, we could
have started the server like:

  t = Tree(..., layouts=[a1, a2, ...])
  t.explore()

with the list of layouts including all the layouts that we may want to
use for a future loaded tree using /load.