average_distance() -> mean_distance().
The name with "mean" instead of "average" is shorter, more precise,
and closer to things like the name MPD (Mean Phylogenetic Distance),
which is what it is when the tree is a phylogenetic tree. And we also
add that detail in the docstring now.

It is a minor change, but it seems like a net gain in all that.