Draw arbitrary faces as headers in the aligned panel.
This makes the headers in the aligned panel very powerful. Any Face
can use position="header" to draw itself in the correct place.

And if there are multiple faces drawn on the same header column, they
will be stacked in order, which shows nicely multiple values in the
same column aligned with those headers too.

The headers are normally used only in the draw_tree function of the
layouts.

For the moment, there are no headers in circular mode. But at least
for rectangular mode they work pretty well.