draw.js: Fine-tune the appearance of headers and aligned items.
Now they never hide information from the first nodes if the first
nodes are easily visible.

Also, there is more space between the dividing line and the first
graphics in the aligned panel, which make it more pleasant.

The headers stay on the top of their columns, always visible if any
part of the tree is visible.