Add TextArrayFace to put headers on top of HeatMapFace and so on.
Now we can have nice texts on top of each little square of data on
heatmaps, for example.

The TextArrayFace is intended mainly to be used from the draw_tree()
function of a layout, in the header position, like all headers.

This took a lot of work to get the anchors right, including when the
text is rotated. I think the changes in the anchoring function
probably help other drawings of text rotated, but I fear something may
have broken somewhere. Not that I found anything broken.