draw.js: Fix order of aligned panels.
Javascrip. Oh, javascript. The keys of objects are strings even if you
put them as numbers. And when you sort them, "10" goes before "2". And
do not get me started with the negative numbers.

Anyway, for a small number of aligned panels, the error was not
visible because the ordering is the same. But I got a big surprise
when going beyond 10, and also when using negatives (which we use to
mean "headers" for the positive aligned panels).

This would be nicer with some more documentation in the code and
possibly elsewhere too.