Fix losing the legend when changing trees or turning on/off layouts.
Otherwise it disappears, because we compare div_legend.innerHTML to
the old one and only replace it if needed (in draw.js), to avoid
flickering.

Maybe it is easier not to have this condition when changing, and
always redraw the legend on calling to update() -> draw(). Before
that, it would be good to check if the flickering is actually there.