contextmenu.js: Make the node_id visualization selectable.
When we were showing the node id, we could not select it and use it
elsewhere (like in a console where we want to use that id).

This is related to gui.css, where for .text we have user-select:none.

This change allows us to show the node id in a way that the user can
select and copy with the mouse, to use later somewhere else.