operations.pyx: Add traverse_full().
This function is like traverse(), but also yields if a node has
already been seen in the traversal (useful for traversing in
prepostorder), and the total distance from the original node to the
current node.

This information is useful for quite a few other functions, even if
not for most traversals. Because of that, and because it is a bit more
complex and slow, I put it as a separate function here.

We will use it in the next couple of commits.