operations.pyx: Simplify farthest_descendant() and change farthest -> farthest_nodes().
Those are two unrelated changes (but it was going to be more messy for
me to put them in separate commits).

Using the previously added traverse_full(), now we can simplify
farthest_descendant() and also make it a bit faster and take less
memory.

And the name of farthest() for the function that returns the two
farthest nodes and their distance (tree diameter) seemed a bit
confusing to me, so I changed it to farthest_nodes().

It is not being used outside for the moment, and there are no tests
related to it. It would be nice to add some.