operations.pyx: Add closest_leaf().
This function uses the newly added traverse_full(), and optimizes
searching by avoiding the traversal of branches that are already
too long to contains the closest leaf.

This function is much faster than t.get_closest_leaf(), and more
flexible.

For speed, it is more than 1000x faster for a tree of ~ 1M leaves.

For flexibility, we can choose which leaves to select, and we can
specify a maximum distance we care about. This will be useful already
in a new function that we will add in the next commit: closest_relative().