operations.pyx: Add closest_relative().
This function finds the closest relative leaf to the given leaf.

It is very fast, and covers many cases.

In particular, for the exercise that we normally do in the ete course,
we could find the closest relative to our isolate, avoiding all
neighboring isolate paralogs, with something like:

  closest_relative(isolate, selector=lambda n: n.name != 'isolate')

This way we do not have to worry about going through monophyletic
branches or not, and we do not have to precompute anything with
t.get_cached_content(), and also the algorithm is correct and finds
the closest even if we have to go much farther into ancestor nodes.

This code could use some better naming and comments, and tests.