draw.py: Fix the style of complex labels.
The name we were using for the label style, when the expression for
the label was complex, did not correspond to the one we used in the
gui (with label.js accessing the label by its class name but using a
different expected name).

This made impossible to change the style of certain labels from the
gui (like color, or font type). For example the one corresponding to
p.get('dist', 0).

This commit fixes all that.