tree.pyx: Add parent and down members. Like up and children.
There are different ways to refer to the connection of the nodes. One
could say:

  parent, children

which are easily understandable family relations. Or:

  up, down

which are positions based on a view of the tree where the root node is
above and the tree grows downwards.

But ete uses none. It mixes inconsistently "up" for "parent", but then
it is "children" and not "down".

This commit introduces the alternative syntax. Now we can choose to
use parent or up, and children or down. All consistent if we want, all
inconsistent if we want.

I would prefer to have just parent and children, but this seems like
the lesser evil.