Fix naming of closest_leaf() and related functions. Implement correctly closest_relative_leaf().
The changes in operations.pyx are:

- Changed closest_leaf() -> closest_descendant_leaf()
- Changed closest_relative() -> closest_leaf() (which also accepts relative=True)
- Added closest_relative_leaf()

The names are more consistent with what they actually do. The closest
leaf is not always a descendant leaf. And the closest *relative* leaf
is not necessarily the closest leaf either.

The closest relative leaf should be within the nearest branch in the
lineage. Since the implementation was going to be very similar to
closest_leaf(), I added instead one argument (relative) that, if True,
computes the closest relative leaf. And the closest_relative_leaf() is
just a convenience function.

And the changes in tree.pyx:

- Removed t.closest_relative() since it had the wrong meaning
- Commented on the meaning of t.get_closest_leaf() and the changes I propose

I recently discussed all these new functions in operations.pyx with
Jaime, and after better understanding I made the changes.

We did not mention those functions in the tutorial, and there are also
no unit tests for them. If we are happy with how they look like now,
it would be nice to add all that.