operations.pyx: Add my own implementation of robinson-foulds.
This is an implementation that needs reviewing, but I fear that if I do not
add it now, it is going to be lost, and I think it is a good implementation
(kind of fast, readable, reasonable arguments and defaults).

In my tests, it seems to work better than t.robinson_foulds() (in most cases
they are the same, but where they differ, it seems to me that the one here
has the correct behavior). To compare them, take into account that this is
the standard robinson-foulds distance for unrooted trees.

Again, I am no expert. This would need review and tests.