operations.pyx: Better estimate the error of NRI and NTI.
When comparing to the tolerance we allow when computing the NRI or the
NTI, take into account the relative error coming from the 1/std term
too.

This is a crude approximation (error_std ~ diff between max and min of
the last nmin std values), but reasonable(?).

In the few tests I have done, it seems to be underestimating the real
error though.

Still, it could be useful approximation.