Allow zoom in aligned panel to be taken into account in the backend.
This commit makes the frontend always send the zoom in the aligned
panel when requesting to draw (view.aligned.zoom, called "za" in the
query parameters).

The backend in explorer.py receives and calls draw() from draw.py not
with a tuple of (zx, zy), but with (zx, zy, za).

Everything is taken into account correctly for the sizes, and then,
only when drawing the faces, it applies the extra zoom of the aligned
panel only for the faces that are going to appear there.

This way, the faces do not need to be modified. They always know the
exact zoom at which they are being visualized.