eval.py: Make sure dist and support are always floats.
If we read support from an extended newick, it will be in node.props
as a string. Yes, there are newicks like that. And when we represent
the support with a PropFace, it will produce an error when formatting
it like "%g" or similar.

This change makes sure this will not happen.

And it fixes the example in ete-gallery examples/seqface.py, which
reads a test tree with that exact problem.