Change the meaning of view.aligned.zoom and of postwidth. Fix zoom in sequences/heatmaps.
This commit changes the meaning of view.aligned.zoom. Now it is not an
extra zoom to add to view.zoom.x, but instead it is directly the zoom
level in the aligned panel.

It also changes the meaning of "postwidth" in SeqFace, HeatmapFace,
and TextArrayFace. Now it is the width of each of the little
squares/positions for zoom=1. And if we change the zoom, everything
will contract or expand on the x axis.