Fix some of the examples.
But the examples in this directory are not great, and the tutorial is
better for these things.

If we wanted good examples, we would need to really review all the
things in the examples/ directory, removing a lot of things, changing
how many things are done, and adding other examples.

This change at least "fixes" some of them, so they work with ete4,
even if they do not illustrate much.