Make SeqFace accept marks.
SeqFace now accepts the argument marks, a list of positions to mark when visualizing the sequence.

This commit requires changes in spritesheet.json and .png too, to include a "marker" rectangle
to use around letters in pixi.

Since I had to change it, I also added the possibility in pixi to draw sequences without text,
which should happen when SeqFace has draw_text=False (in the past, this did work only for svg,
not for pixi, so this commit fixes that bug too).

The way it is done in pixi, it reuses the same png for the colors of the aminoacids, and in
the spritesheet.json it specifies the first pixel of each aminoacid as defining the color
to use when not showing text.