Convert the http server in the explorer to multithreaded.
We now use cheroot multithreaded wsgi server, instead of wsgiref.simple_server.

By default, it creates 100 threads, which is good when we are limited
not by CPU but by I/O (like database access). The number of server
threads can be controlled by the numthreads server parameter.

We can pass extra parameters to the cheroot server with the
server_args argument in start_server() and explore(). For the cheroot
server options, see
https://cheroot.cherrypy.dev/en/stable/pkg/cheroot.wsgi/#cheroot.wsgi.Server

We also changed explorer.stop_server() so it optionally accepts the
thread and server to stop (and now we use it in Tree.render_sm(), for
more clarity). Also, stop_server() is now more symmetric with respect
to start_server().

The logging mechanism in cheroot is very different, so we had to
create a pluging for bottle to log the requests (when we use
t.explore(..., verbose=True) for example). For bottle plugins, see
https://bottlepy.org/docs/dev/plugins/index.html