Add get_server_address() and simplify starting the ete server.
We changed start_server() so that it automatically tracks the created server in g_threads (unless we set track=False).

It also sets the compression (with the compress argument) to use when communicating. So we do not have to do it somewhere else. I think it makes sense here.

Finally, with the new get_server_address() which returns (host, port),
we can simplify quite a bit of code, and make it look independent of
the kind of server that we use (wsgiref.simple_server, cheroot,
whatever).