Small fixes and simplifications for tree ids, /trees/<tree_id> endpoint, layout functions and /load example.
In the distant past, we had tree ids that were numbers of the rows
where they were stored in a sqlite database, and we had translations
from the tree names into those ids. Now we do not use sqlite for
managing trees nor users, and we also use the tree names directly
instead of numeric tree ids.

We have removed GET /trees/<tree_id> , which used to return that translation.

Also, the /trees endpoint now returns a list with the tree names that
the server knows about (instead of a dictionary of correspondences).

In gui.js, the global variable trees now contains an array with the
tree names, instead of an object looking like {x: x}.

There are also changes in the storing of the layouts. g_layouts[None]
are now the layouts stored in the server but not belonging to any
particular tree.

We define the functions get_layouts() and add_layouts() too, for
convenience when working with the ete server.

Finally, we moved the /load endpoint to the example in ete_server.py.

And some small fixes for the gui when no tree is loaded, and cleanup.