Improve t.render_sm().
This includes one fix. render_sm() was using the main server, instead
of the one that it was launching on purpose, so it would not work if
we had not the server already running.

Also, now by default it leaves enough space for the full tree to be
rendered and be quite visible. Of course, big trees make big images.

And it includes many more chrome flags to avoid different annoyances.