draw.js: Fix graphics getting into the header.
Oops. In the previous commit, the white rectangle I removed was
important. We put it back now, with a more precise width now that we
have xmin and xmax too.

When we move the tree up, we want the information on the aligned panel
not to appear on top of the text of the header. This white rectangle
makes a clean space where the text can go afterwards.