Remove unused update_license.py and related.
The update_license.py and FILE_HEADER.txt were not used anywhere.

Also, the license terms are clear, so there seems to be no need for
something like this in the future.

The file release.py is not used either, and is not updated (it still
mentions ete3). But since we might want something like this in the
future, instead of removing it, now it warns about its state, and
mentions the place to look for release information now:
doc/internals/internals_development.rst
(which I will add in a following commit.)