README.md: Remove outdated badges (travis-ci and coveralls.io).
As far as I know, they make no sense now, and have not for a long time.

Not that I think the other badges make a lot of sense, but at least
they seem to work.

I also updated the link to the CODE_OF_CONDUCT.md so it will show well
in the new releases in PyPI, at https://pypi.org/project/ete4/ .